fix(notify): wire AI tool path + prevent window destruction on red-button close
- ai_tools::create/update/delete_schedule now call cmd_notify hooks
  after the SQL mutation, keeping the in-process scheduler in sync with
  DB state. AI-driven writes no longer leak stale notification tasks or
  skip apply_for when the user asks it to adjust a timed reminder.
- WindowEvent::CloseRequested on macOS now prevents the close and hides
  the window, so clicking the Dock icon brings the (still-running)
  window back. The Destroyed path — and its auto-backup + MLX cleanup —
  still runs on real app exit (Cmd+Q).
- Removed the redundant `as i64` cast on i64 in cmd_notify, and deleted
  the dead K_NOTIF_PERMISSION constant (was reserved for a UI cache
  that didn't ship).

+++ src-tauri/src/cmd_notify.rs
@@ -42,7 +42,7 @@ impl ReminderKind {
 pub fn notification_id(schedule_id: i64, kind: ReminderKind) -> i32 {
     // i64 schedule ids above INT_MAX/10 would overflow; in practice the SQLite
     // autoincrement on a personal app stays tiny. Clamp defensively.
-    let base = (schedule_id as i64).min(i32::MAX as i64 / 10) as i32;
+    let base = schedule_id.min(i32::MAX as i64 / 10) as i32;
     base * 10 + kind.code()
 }
